+++ logic/get_user.go
@@ -7,7 +7,7 @@ import (
 	"./../models"
 )
 
-func GetUser(id int64) (user *models.User) {
+func GetUser(id int) (user *models.User) {
 	user = new(models.User)
 
 	err := DB.QueryRow("SELECT id, email, first_name, last_name FROM users WHERE id = $1", id).Scan(&user.Id, &user.Email, &user.FirstName, &user.LastName)
